refactor normal.cpp into parts expression + polynomial manipulation
